### 2018-03-15 #628 @whedon
[PRE REVIEW]: Mapelia and friends: create 3D models from maps
**Submitting author:** @jordibc (<a href="http://orcid.org/0000-0002-9198-5380">Jordi Burguet-Castell</a>)
**Repository:** <a href="https://github.com/jordibc/mapelia" target ="_blank">https://github.com/jordibc/mapelia</a>
**Version:** 1.1.1
