### 2018-03-21 72004fb
add option to put an equator
