### 2018-03-25 391cab5
small speedup at points_at_extreme()
