+++ maps.py
@@ -34,7 +34,7 @@ from numpy import cos, sqrt, pi, array, zeros, average
 
 try:
     import projections as pj
-    assert pj.__version__ == '1.2.0'
+    assert pj.__version__ == '1.2.1'
 except (ImportError, AssertionError, AttributeError) as e:
     sys.exit('projections module not ready. You may want to first run:\n'
              '  %s setup.py build_ext --inplace' % sys.executable)
