+++ projections.c
@@ -788,13 +788,13 @@ struct __pyx_obj_11projections___pyx_scope_struct_5_get_faces {
  * 
  * def points_at_extreme(points, sample_points=[]):             # <<<<<<<<<<<<<<
  *     "Return a list of points that correspond to the boundary of the given ones"
- *     def rxy(p):
+ *     def r2xy(p):
  */
 struct __pyx_obj_11projections___pyx_scope_struct_6_points_at_extreme {
   PyObject_HEAD
   PyObject *__pyx_v_points;
-  PyObject *__pyx_v_rxy;
-  PyObject *__pyx_v_rxy_limit;
+  PyObject *__pyx_v_r2xy;
+  PyObject *__pyx_v_r2xy_limit;
   PyObject *__pyx_v_sample_points;
 };
 
@@ -802,9 +802,9 @@ struct __pyx_obj_11projections___pyx_scope_struct_6_points_at_extreme {
 /* "projections.pyx":336
  * 
  *     sample_points = sample_points or points[1]
- *     rxy_limit = sum(rxy(p) for p in sample_points) / len(sample_points)             # <<<<<<<<<<<<<<
+ *     r2xy_limit = sum(r2xy(p) for p in sample_points) / len(sample_points)             # <<<<<<<<<<<<<<
  * 
- *     return sorted((p for row in points for p in row if rxy(p) > rxy_limit),
+ *     return sorted((p for row in points for p in row if r2xy(p) > r2xy_limit),
  */
 struct __pyx_obj_11projections___pyx_scope_struct_7_genexpr {
   PyObject_HEAD
@@ -817,9 +817,9 @@ struct __pyx_obj_11projections___pyx_scope_struct_7_genexpr {
 
 
 /* "projections.pyx":338
- *     rxy_limit = sum(rxy(p) for p in sample_points) / len(sample_points)
+ *     r2xy_limit = sum(r2xy(p) for p in sample_points) / len(sample_points)
  * 
- *     return sorted((p for row in points for p in row if rxy(p) > rxy_limit),             # <<<<<<<<<<<<<<
+ *     return sorted((p for row in points for p in row if r2xy(p) > r2xy_limit),             # <<<<<<<<<<<<<<
  *                   key=lambda p: arctan2(p[2], p[1]))
  * 
  */
@@ -1357,7 +1357,6 @@ static const char __pyx_k_phi[] = "phi";
 static const char __pyx_k_pid[] = "pid";
 static const char __pyx_k_red[] = "red";
 static const char __pyx_k_row[] = "row";
-static const char __pyx_k_rxy[] = "rxy";
 static const char __pyx_k_sin[] = "sin";
 static const char __pyx_k_sum[] = "sum";
 static const char __pyx_k_args[] = "args";
@@ -1375,11 +1374,12 @@ static const char __pyx_k_norm[] = "norm";
 static const char __pyx_k_nphi[] = "nphi";
 static const char __pyx_k_nx_2[] = "nx_2";
 static const char __pyx_k_ny_2[] = "ny_2";
+static const char __pyx_k_r2xy[] = "r2xy";
 static const char __pyx_k_send[] = "send";
 static const char __pyx_k_sphi[] = "sphi";
 static const char __pyx_k_sqrt[] = "sqrt";
[…]
