### 2018-04-01 #628 @arfon
@hugoledoux @ThomasA @sealhuang - would you be willing to review this submission for JOSS?
