### 2018-04-03 #628 @ThomasA
@arfon yes, I can review this.
