### 2018-04-23 #660 @hugoledoux
I am a GIS (geographical information systems) person used to work with datasets, maps, and 3D data.
So our backgrounds are very different.

## Maps?!

What is a 'map'? That term in GIS is... vague and can mean:
  1. a paper map
  2. a digital version of one, à la Google Maps

You use it to refer to gridded datasets where the value of each pixel is the elevation or something else. I would be a bit more descriptive here.

## Install

I installed it as described, with Python 3.6.5 under macOS. I assume the same would compile under Linux.

I think you should specify that Pillow and numpy are necessary in the install section, I'd have saved a few seconds.

I wonder however if it's possible under Windows?

## version

There is no release, thus I can't confirm that v1.1.1 is there... Releases would really help in knowing what is what, now there's only the master branch.

## A statement of need

    Do the authors clearly state what problems the software is designed to solve and who the target audience is?

That is not entirely clear at this moment. Perhaps a line or two would help? What do you do with the STL files?

## Automated tests

These are not there, but I am not sure if they are needed.

## pintelia

Why is the use of the textures a separate program? Why not just an option of mapelia?

## stl-split

"Split an stl into its north and south sides."
-->
"Split an stl into its north and south HEMISPHERES."

## paper.md

Perhaps the references to projections and formats that are in README.txt should be in here too as references?

## community guidelines

I don't see any file there.
