### 2018-04-23 a4673a1
Be more explicit about what kind of maps we are talking about.
