### 2018-04-23 670b4df
Fix link to Pillow and explain a bit more the prerequisites
