### 2018-04-23 600c5d6
link to tests.py from the root directory
So it is more visible.
