+++ tests.py
@@ -1,0 +1,1 @@
+examples/tests.py
