+++ examples/tests.py
@@ -381,6 +381,7 @@ def modify_ply():
 
 
 if __name__ == '__main__':
+    os.chdir(os.path.dirname(os.path.realpath(__file__)))
     test_mapelia()
     test_pintelia()
     test_poligoniza()
