### 2018-04-23 2acd554
paper.md: add references and explain an existing use-case for the software
