### 2018-04-23 63cd443
readme.rst: fix description of stl-split
