### 2018-04-23 34f0252
readme.rst: add numpy and blender to the references
