### 2018-04-23 cc86814
+ contributing.rst with contributing community guidelines
