+++ contributing.rst
@@ -1,0 +1,21 @@
+Contribute to the software
+--------------------------
+
+If you have an idea or some code that you think would be good to have
+in mapelia, you can just send me a patch to
+`jordi.burguet.castell@gmail.com`.
+
+If you use github, you can also fork the repository and create a pull
+request with your changes. See:
+https://help.github.com/articles/creating-a-pull-request/
+
+Report issues
+-------------
+
+No slack channel, subreddit, irc channel or anything fancy like
+that. Just send me an email (`jordi.burguet.castell@gmail.com`).
+
+Seek support
+------------
+
+You guessed! Pester `jordi.burguet.castell@gmail.com`.
