### 2018-04-23 #660 @jordibc
Thanks @hugoledoux for the review! I found all the comments appropriate and useful. I've tried to address them:

* maps: added clarification (both in readme.rst and paper.md)
* install: added information about the necessary packages in readme.rst
* version: created a release https://github.com/jordibc/mapelia/releases/tag/v1.1.1
* a statement of need: added to paper.md an explanation of how it's used in the project A Touch of The Universe
* automated tests: actually, there were tests, but too hidden in the `examples/` directory, so I've added a link to them (`tests.py`, that can be run directly)
* pintelia: this one is in fact quite a bit different, it does not generate any kind of reliefs (only colored images superposed to a sphere) and only outputs in `ply` (because `stl` does not support colors, even if it is the format commonly used to generate 3D models for printing)
* stl-split: fixed!
* paper.md: good point too, references added
* community guidelines: added `contributing.rst` with the relevant points

Cheers,
J.
