### 2018-04-26 #660 @ThomasA
I have just started reviewing the submission and I will be making remarks here as I go along. I hope that is OK. I may not be able to finish it today...
