### 2018-04-26 #660 @ThomasA
**Authorship**: It is probably as it should be, but I noticed that there are two authors, Amelia Ortiz-Gil and Jordi Burguet-Castell. On the other hand, only Jordi appears to have made [contributions](https://github.com/jordibc/mapelia/graphs/contributors) in the code repository.
