### 2018-04-26 #660 @ThomasA
**Installation instructions**: Just for those interested, the prerequisites can also easily be installed in a new Anaconda environment (here called *mapelia-review*) using:

    conda create -n mapelia-review anaconda=5 pillow numpy

Provided that libgtk is already installed.
