### 2018-04-26 #660 @jordibc
Just FYI @ThomasA , I'm totally fine with doing the remarks as you go along. I'll wait for my reply until you give me a heads up though, unless you prefer it otherwise.
