### 2018-04-27 #660 @ThomasA
@arfon, how do I make it store the ticks on my check list? If I reload the page they are gone...
