### 2018-04-28 #660 @ThomasA
**References**: I am not quite sure what the policy is concerning references, but 'paper.bib' contains two references that do not seem to be included in the paper. On the other hand, the references in the paper are not from the paper.bib file. Maybe this is the way it is intended to be, but in that case I do not see any reason to include 'paper.bib' in the repo.
