### 2018-04-28 #660 @ThomasA
**A statement of need** (documentation): I think that the documentation (in 'readme.rst') could benefit from borrowing a few lines from the paper. The paper starts out with a bit of explanation of what the package is for. I do not find that reflected in the documentation.
