### 2018-04-28 #660 @ThomasA
I think this submission meets the requirements overall and I am ready to recommend acceptance, but I would like to hear feedback on my comments on [authorship](https://github.com/openjournals/joss-reviews/issues/660#issuecomment-384677208), [references](https://github.com/openjournals/joss-reviews/issues/660#issuecomment-385200521), and [a statement of need](https://github.com/openjournals/joss-reviews/issues/660#issuecomment-385200654) first.
