+++ paper/paper.bib
@@ -6,7 +6,7 @@
   urldate = {2018-03-13}
 }
 
-@online{eutils_api_key,
+@online{astrokit,
   author = {Astrokit group},
   title = {Astrokit},
   year = 2018,
