### 2018-04-29 67d48a4
readme.rst: write a statement of need
