### 2018-04-29 #660 @jordibc
Thanks @ThomasA ! Let me try to address those:

* Authorship: you are right, I did write the code, but Amelia got the idea, saw the directions in which it could grow, made many suggestions, tested it extensively and worked with it to generate the actual 3D-printed models --thus why she is an author too.
* References: I didn't realize the references were not used, good catch. I fixed them now. As for the ones included in `paper.md` itself, they provide context that is helpful for the work that is done with the program, but more in the category of details that I didn't consider worth discussing in the paper. Also, I find useful that they are organized in their own sections. But I am not quite sure about the policy there either.
* A statement of need: added to `readme.rst` (actually, now it reads the same as the paper).

All the changes are in the master branch. Please let me know if I should I create a tag. I suppose I should at least create one once the review is done, but I'm not sure.

Thanks also for the tip about anaconda!

Cheers,
J.
